DB Script: Fixed faulty method name extraction.
The problem has appeared when I have removed the '-o' flag from the 'prcegrep' command (due to it duplicating results). This has caused the 'pcregrep' results to include the non-matching line prior to the method block (as the selector included the newline for performance reasons). In some controllers this line included the Route attribute, which was remeniscent of a method signature part, which has caused the subsequent selection (one that's fixed in this commit) to trip.